Fix support for vscode/github.dev
With the addition of the "Current branch" view, the support for vscode.dev/github.dev was inadvertently broken, since I needed more information about the current branch. This mostly skips the "current branch" view for web and re-enables the basic functionality that had been available before.

This is not the right long-term fix, but to work-around the issues for now.