Update development instructions
I've consistently run into issues when the Actions extension is enabled in the development instance of VS Code and you try to test changes in the extension host.  There are errors about duplicate VS Code actions and several other things.  I added a step to the dev instructions to disable the extension.

I also changed the example workspace directory, because this is not the parser repo, and renamed `workspaces` to `workspace` for grammatical correctness